pyproject.toml
[tool.poetry]
name = "my_site"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
django-mptt = "^0.16.0"
django-mptt-admin = "^2.6.2"
django-ckeditor = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
